requirements.txt
click >= 8.0
cloud-files>=4.21.1
requests>=2.25.0
grpcio>=1.36.1
certifi>=2020.12.5
networkx>=2.1
google-cloud-bigtable>=0.33.0
google-cloud-datastore>=1.8
middle-auth-client>=3.11.0
dracopy==1.3.0
zmesh>=1.7.0
fastremap>=1.14.0
jsonschema<4.0
cloud-volume==8.26.0
pyopenssl
flask
numpy
pandas
flask_cors
codecov
multiwrapper
python-json-logger
zstandard
redis
rq
messagingclient
pyyaml
datastoreflex
